feat: Implemented management for menu

+++ RestaurantManagerApp/Utils/ApplicationSettings.cs
@@ -2,7 +2,7 @@
 {
     public class ApplicationSettings
     {
-        public double MenuDiscountPercentageX { get; set; }
+        public decimal MenuDiscountPercentageX { get; set; }
         public double OrderDiscountThresholdY { get; set; }
         public int OrderDiscountMinOrdersZ { get; set; }
         public int OrderDiscountTimeIntervalT_Days { get; set; }
